refactor: Move library metadata to a new meta module
