+++ onvif/__init__.py
@@ -1,9 +1,4 @@
-"""onvif-python: A Python library for interacting with ONVIF-compliant devices."""
-
-from typing import Final
-
-__version__: Final[str] = "0.2.11"
-__repository__: Final[str] = "https://github.com/nirsimetri/onvif-python"
+"""ONVIF Python"""
 
 from onvif.cli import main as ONVIFCLI
 from onvif.client import ONVIFClient
@@ -27,6 +22,4 @@ __all__ = [
     "ONVIFCLI",
     "ONVIFDiscovery",
     "ONVIFParser",
-    "__version__",
-    "__repository__",
 ]
