+++ onvif/meta.py
@@ -1,0 +1,6 @@
+"""onvif-python: A Python library for interacting with ONVIF-compliant devices."""
+
+from typing import Final
+
+__version__: Final[str] = "0.2.11"
+__repository__: Final[str] = "https://github.com/nirsimetri/onvif-python"
